Provide Confidence Interval for Provider effects or Standardization Ratios/Rates

Provide Confidence Interval for Provider effects or Standardization Ratios/Rates

# S3 method for logis_fe
confint(
  fit,
  parm,
  level = 0.95,
  test = "exact",
  option = c("gamma", "SR"),
  stdz = "indirect",
  measure = c("rate", "ratio")
)

Arguments

fit

an object as output of logis_fe function

parm

specify a subset of which providers are to be given confidence intervals.

level

confidence level used for constructing confidence intervals. Defaulting to 0.95

test

a character string specifying the type of testing method, defaulting to "exact" * `exact`: exact test * `wald`: wald test * `score`: score test

option

the confidence interval for the function's output, whether it is for gamma or standardization ratios/rates * `gamma`: provider effect * `SR`: standardization ratios/rates

stdz

if option = 'SR', a character string specifying the standardization method. Defaulting to "indirect". * `indirect`: using indirect standardized method * `direct`: using direct standardized method

measure

if option = 'SR', a boolean indicating whether the output measure is "ratio" or "rate". Both "rate" and "ratio" will be provided by default. * `rate`: output the standardized rate. The "rate" has been restricted to 0 * `ratio`: output the standardized ratio

...

Value

A dataframe containing the point estimate, and lower and upper bounds of the estimate.

Examples

data(data_FE)
data.prep <- fe_data_prep(data_FE$Y, data_FE$Z, data_FE$ID)
#> Checking missingness of variables ... 
#> Missing values NOT found. Checking missingness of variables completed!
#> Checking variation in covariates ... 
#> Checking variation in covariates completed!
#> Checking pairwise correlation among covariates ... 
#> Checking pairwise correlation among covariates completed!
#> Checking VIF of covariates ... 
#> Checking VIF of covariates completed!
#> Warning: 0 out of 10 providers considered small and filtered out!
#> 0 out of 10 remaining providers with no readmission within 30 days.
#> 0 out of 10 remaining providers with all readmissions within 30 days.
#> After screening, 52% of all discharges were readmitted within 30 days.
fit_fe <- logis_fe(data.prep)
confint(fit_fe, option = "gamma")
#>           gamma gamma.lower gamma.upper
#> 1  -0.053509017  -0.9269014   0.8208674
#> 2   0.772595018  -0.1861332   1.7290101
#> 3  -1.389717213  -2.3996017  -0.4379263
#> 4   2.517923014   1.4561906   3.5755053
#> 5   0.272975840  -0.7196352   1.2802052
#> 6  -0.436788388  -1.4678909   0.5411435
#> 7   0.008545775  -0.9158587   0.9000023
#> 8   0.304527080  -0.7020479   1.3619158
#> 9   1.727970674   0.6661555   2.7870210
#> 10 -0.034881361  -0.9473625   0.8875052
confint(fit_fe, option = "SR", stdz = "direct", measure = "rate")
#> $CI.direct_rate
#>    direct_rate CI_rate.lower CI_rate.upper
#> 1     49.27094      41.72223      56.88122
#> 2     56.46127      48.11855      64.72792
#> 3     37.79631      29.60944      45.93492
#> 4     71.31486      62.38719      79.32001
#> 5     52.11158      43.50198      60.86772
#> 6     45.94477      37.14172      54.44656
#> 7     49.81051      41.81677      57.56939
#> 8     52.38627      43.65346      61.57397
#> 9     64.71905      55.53493      73.46444
#> 10    49.43288      41.54713      57.46074
#>