Main function for fitting fixed effects model

Main function for fitting fixed effects model

Usage

logis_fe(
  data.prep,
  algorithm = "SerBIN",
  max.iter = 10000,
  tol = 1e-05,
  bound = 10,
  backtrack = TRUE,
  Rcpp = TRUE,
  threads = 1,
  AUC = FALSE,
  message = FALSE
)

Arguments

data.prep

an object from the fe_data_prep() function.

algorithm

a string specifying the algorithm to be used. Defaulting to "SerBIN".

"SerBIN": using the Serial blockwise inversion Newton algorithm to fit the model (See Wu et al. (2022)).

"BAN": using the block ascent Newton algorithm to fit the model (See He et al. (2013)).

max.iter

maximum number of iterations. Defaulting to 10,000.

tol

a small positive number specifying stopping criterion of Newton-Raphson algorithm. Defaulting to 1e-5.

bound

a positive number to avoid inflation of provider effect. Defaulting to 10.

backtrack

a boolean indicating whether backtracking line search is implemented. Defaulting to FALSE.

Rcpp

a Boolean indicating whether the Rcpp function is used if "Rcpp = T". Defaulting to TRUE.

threads

a positive integer specifying the number of threads to be used. Defaulting to 1.

AUC

a Boolean indicating whether report AUC. Defaulting to FALSE.

message

a Boolean indicating whether track the fitting process. Defaulting to TRUE.

...

Value

An object with S3 class logis_fe.

beta

a vector of fixed effects estimates of covariates

gamma

a vector of estimates of provider effects

linear_pred

a vector of linear predictors

pred

a vector of predicted probabilities

neg2Loglkd

minus two times log likelihood

AIC

Akaike info criterion

BIC

Bayesian info criterion

AUC

area under the ROC curve

Details

The default algorithm is based on Serial blockwise inversion Newton (SerBIN) proposed by Wu et al. (2022), but users can also choose to use the block ascent Newton (BAN) algorithm proposed by He et al. (2013) to fit the model. Both methodologies build upon the Newton-Raphson method, yet SerBIN simultaneously updates both the provider effect and covariate coefficient. This concurrent update necessitates the inversion of the complete information matrix at each iteration. In contrast, BAN adopts a two-layer updating approach, where the covariate coefficient is sequentially fixed to update the provider effect, followed by fixing the provider effect to update the covariate coefficient.

We suggest using the default "SerBIN" option as it typically converges much faster for most datasets. However, in rare cases where the SerBIN algorithm encounters second-order derivative irreversibility leading to an error, users can consider using the "BAN" option as an alternative.

For a deeper understanding, please consult the original article for detailed insights.

References

Wu, W, Yang, Y, Kang, J, He, K. (2022) Improving large-scale estimation and inference for profiling health care providers. Statistics in Medicine, 41(15): 2840-2853.

He K, Kalbfleisch, J, Li, Y, and et al. (2013) Evaluating hospital readmission rates in dialysis providers; adjusting for hospital effects. Lifetime Data Analysis, 19: 490-512.

Examples

data(data_FE)
data.prep <- fe_data_prep(data_FE$Y, data_FE$Z, data_FE$ID)
#> Checking missingness of variables ... 
#> Missing values NOT found. Checking missingness of variables completed!
#> Checking variation in covariates ... 
#> Checking variation in covariates completed!
#> Checking pairwise correlation among covariates ... 
#> Checking pairwise correlation among covariates completed!
#> Checking VIF of covariates ... 
#> Checking VIF of covariates completed!
#> Warning: 0 out of 10 providers considered small and filtered out!
#> 0 out of 10 remaining providers with no events.
#> 0 out of 10 remaining providers with all events.
#> After screening, 52% of all records exhibit occurrences of events (Y = 1)
fit_fe <- logis_fe(data.prep)