Main function for fitting fixed effects model with firth's bias correction

Main function for fitting fixed effects model with firth's bias correction

Usage

logis_firth(
  data.prep,
  max.iter = 10000,
  tol = 1e-05,
  bound = 10,
  backtrack = TRUE,
  Rcpp = TRUE,
  threads = 1,
  AUC = FALSE,
  message = FALSE
)

Arguments

data.prep

an object from the fe_data_prep() function.

max.iter

maximum number of iterations. Defaulting to 10,000.

tol

a small positive number specifying stopping criterion of Newton-Raphson algorithm. Defaulting to 1e-5.

bound

a positive number to avoid inflation of provider effect. Defaulting to 10.

backtrack

a boolean indicating whether backtracking line search is implemented. Defaulting to FALSE.

Rcpp

a Boolean indicating whether the Rcpp function is used. Defaulting to TRUE.

threads

an integer indicating the number of threads to use if "Rcpp = T". Defaulting to 1.

AUC

a Boolean indicating whether report AUC. Defaulting to FALSE.

message

a Boolean indicating whether track the fitting process. Defaulting to TRUE.

...

Value

An object with S3 class logis_fe.

beta

a vector of fixed effects estimates of covariates

gamma

a vector of estimates of provider effects

obs

a vector of patients-level outcome

neg2Loglkd

minus two times log likelihood

AIC

Akaike info criterion

BIC

Bayesian info criterion

AUC

area under the ROC curve

Examples

data(data_FE)
data.prep <- fe_data_prep(data_FE$Y, data_FE$Z, data_FE$ID)
#> Checking missingness of variables ... 
#> Missing values NOT found. Checking missingness of variables completed!
#> Checking variation in covariates ... 
#> Checking variation in covariates completed!
#> Checking pairwise correlation among covariates ... 
#> Checking pairwise correlation among covariates completed!
#> Checking VIF of covariates ... 
#> Checking VIF of covariates completed!
#> Warning: 0 out of 10 providers considered small and filtered out!
#> 0 out of 10 remaining providers with no events.
#> 0 out of 10 remaining providers with all events.
#> After screening, 52% of all records exhibit occurrences of events (Y = 1)
fit_firth <- logis_firth(data.prep)