Provide the summary statistics for covariate estimates

Provide the summary statistics for covariate estimates

Usage

summary_fe_covar(fit, parm, level = 0.95, test = "wald.cpp", null = 0)

Arguments

fit

an object as output of logis_fe function.

parm

a character vector specifies a subset of covariates. All covariates are included by default.

level

confidence level used for constructing confidence intervals. Defaulting to 0.95.

test

a character string specifying the type of testing method. Defaulting to "wald.cpp".

"wald": wald test

"wald.cpp": wald test using cpp function

"lr": likelihood ratio test

"score": score test

null

the null value of the covariate estimate that requires testing. (e.g. test \(H_0: \beta = 0\))

...

Value

a dataframe containing summary statistics for covariate estimates

Examples

data(data_FE)
data.prep <- fe_data_prep(data_FE$Y, data_FE$Z, data_FE$ID, message = FALSE)
fit_fe <- logis_fe(data.prep)
summary.wald.cpp <- summary_fe_covar(fit_fe, level = 0.05, test = "wald")
summary.wald.cpp
#>          beta   se.beta            p   CI.lower  CI.upper
#> Z1 -2.2958595 0.2559326 0.000000e+00 -2.3119082 -2.279811
#> Z2 -1.2672733 0.1927867 4.916267e-11 -1.2793624 -1.255184
#> Z3  0.5674081 0.1744281 1.142086e-03  0.5564703  0.578346
#> Z4  3.4980033 0.3371212 3.185274e-25  3.4768635  3.519143