Hypothesis Tests for Identifying Outlying Providers

Hypothesis Tests for Identifying Outlying Providers

test_fe(
  fit,
  parm,
  level = 0.95,
  test = "exact.poisbinom",
  null = "median",
  n = 10000
)

Arguments

fit

an object as output of logis_fe function.

parm

specify a subset of which providers are to be given confidence intervals. All providers are included by default.

level

confidence level used for constructing confidence intervals. Defaulting to 0.95.

test

a character string specifying the type of testing method to be conducted. Defaulting to "exact.poisbinom".

"exact.poisbinom": two-sided exact test based on Poisson-binomial distribution of \(O_i|Z_i\)

"exact.bootstrap": two-sided exact test based on bootstrap procedure

"wald": wald test

"score": score test

null

a character string or real number specifying null hypotheses of fixed provider effects.

n

resample size for bootstrapping. Defaulting to 10,000.

...

Value

a dataframe containing:

flag

a vector of flagging indicator. "1" means better, and "-1" means worse

p

p-value

stat

z-score

Details

By default, the function uses the "exact.poisbinom" method. The wald test is invalid for extreme providers (i.e. when provider effect goes to infinity).

References

Wu, W, Yang, Y, Kang, J, He, K. (2022) Improving large-scale estimation and inference for profiling health care providers. Statistics in Medicine, 41(15): 2840-2853.

Examples

data(data_FE)
data.prep <- fe_data_prep(data_FE$Y, data_FE$Z, data_FE$ID, message = FALSE)
fit_fe <- logis_fe(data.prep)
test_fe(fit_fe)
#>    flag            p       stat
#> 1     0 6.624388e-01 -0.4365486
#> 2     0 1.965498e-01  1.2914439
#> 3    -1 1.264412e-03 -3.2239370
#> 4     1 1.547222e-05  4.3218560
#> 5     0 7.992864e-01  0.2542707
#> 6     0 2.577034e-01 -1.1318361
#> 7     0 7.828967e-01 -0.2755461
#> 8     0 7.678697e-01  0.2951626
#> 9     1 3.317876e-03  2.9366413
#> 10    0 7.044033e-01 -0.3793832